requirements.txt
google-genai
python-dotenv
gradio
